move types to upcshared project, convert upcshared project into .netstandard20 library
