+++ DroidUpc/DroidUpc.csproj
@@ -67,7 +67,6 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="scanner.cs" />
     <Compile Include="StatusBox.cs" />
-    <Compile Include="types.cs" />
     <Compile Include="UpcAlert.cs" />
     <Compile Include="UpcInvCore.cs" />
     <Compile Include="UpcTypeSpinnerAdapter.cs" />
@@ -152,8 +151,8 @@
     <PackageReference Include="Newtonsoft.Json">
       <Version>12.0.2</Version>
     </PackageReference>
-    <PackageReference Include="TCore.Standard.StatusResult">
-      <Version>1.0.0</Version>
+    <PackageReference Include="TCore.StatusResult">
+      <Version>2.0.0</Version>
     </PackageReference>
     <PackageReference Include="Xamarin.Android.Support.Design" Version="27.0.2.1" />
     <PackageReference Include="Xamarin.Android.Support.v4">
@@ -194,6 +193,10 @@
     <ProjectReference Include="..\..\liblet\cs\TCore.Logging\Logging.csproj">
       <Project>{ee681c45-8416-4de7-9e62-ae1371a7de9a}</Project>
       <Name>Logging</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\UpcShared\UpcShared.csproj">
+      <Project>{85c5de2b-7d29-4744-9406-98858dbc5b33}</Project>
+      <Name>UpcShared</Name>
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
