+++ UpcApi/packages.config
@@ -9,5 +9,5 @@
   <package id="NUnit" version="3.12.0" targetFramework="net472" />
   <package id="TCore.Exceptions" version="1.0.0" targetFramework="net472" />
   <package id="TCore.Sql" version="1.1.0" targetFramework="net472" />
-  <package id="TCore.StatusResult" version="1.0.0" targetFramework="net472" />
+  <package id="TCore.StatusResult" version="2.0.0" targetFramework="net472" />
 </packages>
