+++ UpcInv.sln
@@ -35,6 +35,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bulker", "Bulker\Bulker.csp
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UpcApi", "UpcApi\UpcApi.csproj", "{91D7AD8B-4646-435F-BA19-B1E089D74D41}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UpcShared", "UpcShared\UpcShared.csproj", "{85C5DE2B-7D29-4744-9406-98858DBC5B33}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		AzureRelease|Any CPU = AzureRelease|Any CPU
@@ -793,6 +795,62 @@ Global
 		{91D7AD8B-4646-435F-BA19-B1E089D74D41}.Release|x64.Build.0 = Release|Any CPU
 		{91D7AD8B-4646-435F-BA19-B1E089D74D41}.Release|x86.ActiveCfg = Release|Any CPU
 		{91D7AD8B-4646-435F-BA19-B1E089D74D41}.Release|x86.Build.0 = Release|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.AzureRelease|Any CPU.ActiveCfg = Release|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.AzureRelease|Any CPU.Build.0 = Release|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.AzureRelease|ARM.ActiveCfg = Release|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.AzureRelease|ARM.Build.0 = Release|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.AzureRelease|x64.ActiveCfg = Release|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.AzureRelease|x64.Build.0 = Release|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.AzureRelease|x86.ActiveCfg = Release|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.AzureRelease|x86.Build.0 = Release|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.AzureStagingDebug|Any CPU.ActiveCfg = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.AzureStagingDebug|Any CPU.Build.0 = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.AzureStagingDebug|ARM.ActiveCfg = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.AzureStagingDebug|ARM.Build.0 = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.AzureStagingDebug|x64.ActiveCfg = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.AzureStagingDebug|x64.Build.0 = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.AzureStagingDebug|x86.ActiveCfg = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.AzureStagingDebug|x86.Build.0 = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.Debug|ARM.Build.0 = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.Debug|x64.Build.0 = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.Debug|x86.Build.0 = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.DebugStageHostStageSvcProdData|Any CPU.ActiveCfg = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.DebugStageHostStageSvcProdData|Any CPU.Build.0 = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.DebugStageHostStageSvcProdData|ARM.ActiveCfg = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.DebugStageHostStageSvcProdData|ARM.Build.0 = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.DebugStageHostStageSvcProdData|x64.ActiveCfg = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.DebugStageHostStageSvcProdData|x64.Build.0 = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.DebugStageHostStageSvcProdData|x86.ActiveCfg = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.DebugStageHostStageSvcProdData|x86.Build.0 = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.LocalDebug|Any CPU.ActiveCfg = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.LocalDebug|Any CPU.Build.0 = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.LocalDebug|ARM.ActiveCfg = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.LocalDebug|ARM.Build.0 = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.LocalDebug|x64.ActiveCfg = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.LocalDebug|x64.Build.0 = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.LocalDebug|x86.ActiveCfg = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.LocalDebug|x86.Build.0 = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.LocalStaging|Any CPU.ActiveCfg = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.LocalStaging|Any CPU.Build.0 = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.LocalStaging|ARM.ActiveCfg = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.LocalStaging|ARM.Build.0 = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.LocalStaging|x64.ActiveCfg = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.LocalStaging|x64.Build.0 = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.LocalStaging|x86.ActiveCfg = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.LocalStaging|x86.Build.0 = Debug|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.Release|Any CPU.Build.0 = Release|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.Release|ARM.ActiveCfg = Release|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.Release|ARM.Build.0 = Release|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.Release|x64.ActiveCfg = Release|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.Release|x64.Build.0 = Release|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.Release|x86.ActiveCfg = Release|Any CPU
+		{85C5DE2B-7D29-4744-9406-98858DBC5B33}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
